Move 'WaitForDeployment()' into 'client' package
