+++ cli/command/service.go
@@ -4,6 +4,7 @@ import (
 	"fmt"
 	"strconv"
 
+	"github.com/quintilesims/layer0/client"
 	"github.com/quintilesims/layer0/common/config"
 	"github.com/quintilesims/layer0/common/models"
 	"github.com/urfave/cli"
@@ -126,7 +127,7 @@ func (s *ServiceCommand) create(c *cli.Context) error {
 	}
 
 	onCompleteFN := func(serviceID string) error {
-		service, err := WaitForDeployment(s.client, serviceID, c.GlobalDuration(config.FLAG_TIMEOUT))
+		service, err := client.WaitForDeployment(s.client, serviceID, c.GlobalDuration(config.FLAG_TIMEOUT))
 		if err != nil {
 			return err
 		}
@@ -226,7 +227,7 @@ func (s *ServiceCommand) scale(c *cli.Context) error {
 	}
 
 	onCompleteFN := func(serviceID string) error {
-		service, err := WaitForDeployment(s.client, serviceID, c.GlobalDuration(config.FLAG_TIMEOUT))
+		service, err := client.WaitForDeployment(s.client, serviceID, c.GlobalDuration(config.FLAG_TIMEOUT))
 		if err != nil {
 			return err
 		}
@@ -264,7 +265,7 @@ func (s *ServiceCommand) update(c *cli.Context) error {
 	}
 
 	onCompleteFN := func(serviceID string) error {
-		service, err := WaitForDeployment(s.client, serviceID, c.GlobalDuration(config.FLAG_TIMEOUT))
+		service, err := client.WaitForDeployment(s.client, serviceID, c.GlobalDuration(config.FLAG_TIMEOUT))
 		if err != nil {
 			return err
 		}
